backend/requirements.txt
fastapi
uvicorn
tensorflow-cpu
Pillow
websockets
numpy
python-multipart
opencv-python
python-dotenv